+++ model/subnet_response.h
@@ -24,8 +24,8 @@ typedef struct subnet_response_t {
     int fki_user_id; //numeric
     int fki_apikey_id; //numeric
     struct multilingual_subnet_description_t *obj_subnet_description; //model
-    int i_subnet_network; //numeric
-    int i_subnet_mask; //numeric
+    long i_subnet_network; //numeric
+    long i_subnet_mask; //numeric
 
 } subnet_response_t;
 
@@ -34,8 +34,8 @@ subnet_response_t *subnet_response_create(
     int fki_user_id,
     int fki_apikey_id,
     multilingual_subnet_description_t *obj_subnet_description,
-    int i_subnet_network,
-    int i_subnet_mask
+    long i_subnet_network,
+    long i_subnet_mask
 );
 
 void subnet_response_free(subnet_response_t *subnet_response);
